Fix SyncClient response handling
Firstly, createRepo was broken because it's response type is undefined
(no JSON is returned by PUT /v2/store/<syncKey>).
Secondly, we use cleaners to validate the response JSON for better
error handling and type checking.
Lastly, we add `method` to error response because this is useful info
for v2 API.